move alkan 45 back to pieces and update pdf
